Adjust checks and functions to IPv6 changes. Move Proxy init functions to ProxySetup.go
