+++ node-net-manager/TableEntryCache/serviceTranslationTable.go
@@ -59,14 +59,16 @@ func (t *TableManager) Add(entry TableEntry) error {
 	return errors.New("InvalidEntry")
 }
 
+// remove by Namespace IP, which can be either in IPv4 or IPv6 format
 func (t *TableManager) RemoveByNsip(nsip net.IP) error {
 
 	t.rwlock.Lock()
 	defer t.rwlock.Unlock()
 
 	found := -1
+	// this will need to be optimised for IPv6, since that will be hell performance wise
 	for i, tableElement := range t.translationTable {
-		if tableElement.Nsip.Equal(nsip) {
+		if tableElement.Nsip.Equal(nsip) || tableElement.Nsipv6.Equal(nsip) {
 			found = i
 			break
 		}
@@ -110,7 +112,7 @@ func (t *TableManager) SearchByServiceIP(ip net.IP) []TableEntry {
 	defer t.rwlock.Unlock()
 	for _, tableElement := range t.translationTable {
 		for _, elemip := range tableElement.ServiceIP {
-			if elemip.Address.Equal(ip) {
+			if elemip.Address.Equal(ip) || elemip.Address_v6.Equal(ip) {
 				returnEntry := tableElement
 				result = append(result, returnEntry)
 			}
@@ -123,7 +125,7 @@ func (t *TableManager) SearchByNsIP(ip net.IP) (TableEntry, bool) {
 	t.rwlock.Lock()
 	defer t.rwlock.Unlock()
 	for _, tableElement := range t.translationTable {
-		if tableElement.Nsip.Equal(ip) {
+		if tableElement.Nsip.Equal(ip) || tableElement.Nsipv6.Equal(ip) {
 			returnEntry := tableElement
 			return returnEntry, true
 		}
@@ -143,7 +145,7 @@ func (t *TableManager) SearchByJobName(jobname string) []TableEntry {
 	return results
 }
 
-// Sanity chceck for Appname and namespace
+// Sanity check for Appname and namespace
 // 0<len(Appname)<11
 // 0<len(Appns)<11
 // 0<len(Servicename)<11
@@ -152,22 +154,23 @@ func (t *TableManager) SearchByJobName(jobname string) []TableEntry {
 // Cluster>0
 // Nodeip != nil
 // Nsip != nil
+// Nsipv6 != nil
 // len(entry.ServiceIP)>0
 func (t *TableManager) isValid(entry TableEntry) bool {
 	if l := len(entry.Appname); l < 1 || l > 10 {
-		log.Println("TranslationTable: Invalid Entry, wrong appname")
+		log.Println("TranslationTable: Invalid Entry, wrong appname:", entry.Appname)
 		return false
 	}
 	if l := len(entry.Appns); l < 1 || l > 10 {
-		log.Println("TranslationTable: Invalid Entry, wrong appns")
+		log.Println("TranslationTable: Invalid Entry, wrong appns:", entry.Appns)
 		return false
 	}
 	if l := len(entry.Servicename); l < 1 || l > 10 {
-		log.Println("TranslationTable: Invalid Entry, wrong servicename")
+		log.Println("TranslationTable: Invalid Entry, wrong servicename:", entry.Servicename)
 		return false
 	}
 	if l := len(entry.Servicenamespace); l < 1 || l > 10 {
-		log.Println("TranslationTable: Invalid Entry, wrong servicens")
+		log.Println("TranslationTable: Invalid Entry, wrong servicens:", entry.Servicenamespace)
 		return false
 	}
 	if entry.Instancenumber < 0 {
@@ -186,6 +189,10 @@ func (t *TableManager) isValid(entry TableEntry) bool {
 		log.Println("TranslationTable: Invalid Entry, wrong nsip")
 		return false
 	}
+	if entry.Nsipv6 == nil {
+		log.Println("TranslationTable: Invalid Entry, wrong nsipv6")
+		return false
+	}
 	if len(entry.ServiceIP) < 1 {
 		log.Println("TranslationTable: Invalid Entry, wrong serviceip")
 		return false
@@ -195,7 +202,7 @@ func (t *TableManager) isValid(entry TableEntry) bool {
 
 func IsNamespaceStillValid(nsip net.IP, table *[]TableEntry) bool {
 	for _, entry := range *table {
-		if entry.Nsip.Equal(nsip) {
+		if entry.Nsip.Equal(nsip) || entry.Nsipv6.Equal(nsip) {
 			return true
 		}
 	}
